backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings

langchain
langchain-openai
langchain-huggingface
langchain-groq

chromadb
beautifulsoup4
requests

python-multipart
python-dotenv

aiofiles
pillow
pdfplumber
pytesseract

transformers
sentencepiece